refactor: RouteChecker에게 지나갈 수 있는 경로 확인 책임 넘기기
- Board가 판단하던 것을 RouteChecker에게 위임
